wip move copy mec
