+++ crates/cubecl-linalg/src/matmul/components/global/mod.rs
@@ -6,6 +6,7 @@ pub mod tensor_view;
 mod accumulator_loader;
 mod base;
 mod config;
+mod copy_mechanism;
 mod tilewise_unloading;
 
 pub mod output_loader;
@@ -13,3 +14,4 @@ pub mod output_loader;
 pub use accumulator_loader::*;
 pub use base::*;
 pub use config::*;
+pub use copy_mechanism::*;
