+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/loading/async/mod.rs
@@ -1,0 +1,3 @@
+mod maximize_slice_length;
+
+pub use maximize_slice_length::*;
