+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/loading/mod.rs
@@ -1,5 +1,7 @@
 mod base;
 mod sync;
+mod r#async;
 
 pub use base::*;
 pub use sync::*;
+pub use r#async::*;
