+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/async/maximize_unit_count.rs
@@ -9,7 +9,7 @@ use crate::matmul::components::{
 use cubecl_core::prelude::*;
 use cubecl_core::{self as cubecl, prelude::barrier::BarrierLevel};
 
-use super::{AsyncFullLoadingStrategy, CopyMechanism};
+use super::AsyncFullLoadingStrategy;
 
 #[derive(CubeType, Clone, Copy)]
 /// Executes one memcpy_async call per unit.
